Fix backend unit tests, resolve missing rate limiter, 404, and routing issues

+++ backend/src/entities/deal.rs
@@ -14,7 +14,7 @@ pub struct Model {
     pub id: Uuid,
     pub customer_id: Uuid,  // Reference to the customer
     pub name: String,
-    pub amount: f32,  // Deal amount
+    pub amount: f64,  // Deal amount
     pub status: String,  // e.g., "Prospecting", "Qualification", "Closed Won", "Closed Lost"
     pub stage: String,  // Current stage in the sales process
     pub close_date: Option<DateTime<Utc>>,  // Expected or actual close date
